requirements.txt
click==8.0.1
environs==9.3.3
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Werkzeug==2.0.1
